refactor AsyncAvatarImage
