+++ OctoTrack/ExtractedViews/AsyncAvatarImage.swift
@@ -8,7 +8,6 @@
 import SwiftUI
 
 struct AsyncAvatarImage: View {
-    @State private var viewModel = AsyncAvatarViewModel()
     @State private var image: UIImage = UIImage()
     let avatarName: String
     let avatarUrl: String
@@ -21,7 +20,22 @@ struct AsyncAvatarImage: View {
             .frame(width: size, height: size)
             .clipShape(Circle())
             .task {
-                image = await viewModel.loadImage(named: avatarName, urlString: avatarUrl)
+                image = await loadImage(named: avatarName, urlString: avatarUrl)
             }
+    }
+    
+    private func loadImage(named name: String, urlString: String?) async -> UIImage {
+        if let localImage = UIImage(named: name) {
+            return localImage
+        }
+
+        if let urlString = urlString,
+              let url = URL(string: urlString),
+              let (data, _) = try? await URLSession.shared.data(from: url),
+              let downloadedImage = UIImage(data: data) {
+               return downloadedImage
+           }
+
+        return UIImage(named: "AppLogo") ?? UIImage()
     }
 }
